+++ testing/mochitest/runtestsb2g.py
@@ -9,6 +9,7 @@ import sys
 import tempfile
 import time
 import urllib
+import traceback
 
 sys.path.insert(0, os.path.abspath(os.path.realpath(os.path.dirname(sys.argv[0]))))
 
@@ -40,10 +41,15 @@ class B2GOptions(MochitestOptions):
                     help = "host:port to use when connecting to Marionette")
         defaults["marionette"] = None
 
-        self.add_option("--emulator", action="store_true",
-                    dest = "emulator",
-                    help = "True if using a b2g emulator")
-        defaults["emulator"] = False
+        self.add_option("--emulator", action="store",
+                    type="string", dest = "emulator",
+                    help = "Architecture of emulator to use: x86 or arm")
+        defaults["emulator"] = None
+
+        self.add_option("--no-window", action="store_true",
+                    dest = "noWindow",
+                    help = "Pass --no-window to the emulator")
+        defaults["noWindow"] = False
 
         self.add_option("--adbpath", action="store",
                     type = "string", dest = "adbPath",
@@ -90,6 +96,7 @@ class B2GOptions(MochitestOptions):
         defaults["autorun"] = True
         defaults["closeWhenDone"] = True
         defaults["testPath"] = ""
+        defaults["extensionsToExclude"] = ["specialpowers"]
 
         self.set_defaults(**defaults)
 
@@ -143,7 +150,7 @@ class B2GOptions(MochitestOptions):
         options.sslPort = tempSSL
         options.httpPort = tempPort
 
-        return options 
+        return options
 
 
 class ProfileConfigParser(ConfigParser.RawConfigParser):
@@ -177,6 +184,7 @@ class B2GMochitest(Mochitest):
     _automation = None
     _dm = None
     localProfile = None
+    testDir = '/data/local/tests'
 
     def __init__(self, automation, devmgr, options):
         self._automation = automation
@@ -186,20 +194,36 @@ class B2GMochitest(Mochitest):
         self.remoteProfile = options.remoteTestRoot + '/profile'
         self._automation.setRemoteProfile(self.remoteProfile)
         self.remoteLog = options.remoteLogFile
-        self.remoteProfilesIniPath = '/data/b2g/mozilla/profiles.ini'
+        self.userJS = '/data/local/user.js'
+        self.remoteMozillaPath = '/data/b2g/mozilla'
+        self.remoteProfilesIniPath = os.path.join(self.remoteMozillaPath, 'profiles.ini')
         self.originalProfilesIni = None
 
     def cleanup(self, manifest, options):
-        self._dm.getFile(self.remoteLog, self.localLog)
-        self._dm.removeFile(self.remoteLog)
-        self._dm.removeDir(self.remoteProfile)
-
+        # Restore the original profiles.ini.
         if self.originalProfilesIni:
             try:
-                self.restoreProfilesIni()
+                if not options.emulator:
+                    self.restoreProfilesIni()
                 os.remove(self.originalProfilesIni)
             except:
                 pass
+
+        if not options.emulator:
+            self._dm.getFile(self.remoteLog, self.localLog)
+            self._dm.removeFile(self.remoteLog)
+            self._dm.removeDir(self.remoteProfile)
+
+            # Restore the original user.js.
+            self._dm.checkCmdAs(['shell', 'rm', '-f', self.userJS])
+            if self._dm.useDDCopy:
+                self._dm.checkCmdAs(['shell', 'dd', 'if=%s.orig' % self.userJS, 'of=%s' % self.userJS])
+            else:
+                self._dm.checkCmdAs(['shell', 'cp', '%s.orig' % self.userJS, self.userJS])
+
+            # We've restored the original profile, so reboot the device so that
+            # it gets picked up.
+            self._automation.rebootDevice()
 
         if options.pidFile != "":
             try:
@@ -207,10 +231,6 @@ class B2GMochitest(Mochitest):
                 os.remove(options.pidFile + ".xpcshell.pid")
             except:
                 print "Warning: cleaning up pidfile '%s' was unsuccessful from the test harness" % options.pidFile
-
-        # We've restored the original profile, so reboot the device so that
-        # it gets picked up.
-        self._automation.rebootDevice()
 
     def findPath(self, paths, filename = None):
         for path in paths:
@@ -278,12 +298,12 @@ class B2GMochitest(Mochitest):
         options.profilePath = remoteProfilePath
 
     def stopWebServer(self, options):
-        self.server.stop()
+        if hasattr(self, 'server'):
+            self.server.stop()
 
     def buildProfile(self, options):
         if self.localProfile:
             options.profilePath = self.localProfile
-        print 'buildProfile', repr(options)
         manifest = Mochitest.buildProfile(self, options)
         self.localProfile = options.profilePath
 
@@ -318,6 +338,10 @@ class B2GMochitest(Mochitest):
             config.write(configfile)
 
         self._dm.pushFile(newProfilesIni, self.remoteProfilesIniPath)
+        try:
+            os.remove(newProfilesIni)
+        except:
+            pass
 
     def buildURLOptions(self, options, env):
         self.localLog = options.logFile
@@ -331,15 +355,32 @@ class B2GMochitest(Mochitest):
             testURL += "?" + "&".join(self.urlOpts)
         self._automation.testURL = testURL
 
-        # Set the B2G homepage as a static local page, since wi-fi generally
-        # isn't available as soon as the device boots.
+        # Set extra prefs for B2G.
         f = open(os.path.join(options.profilePath, "user.js"), "a")
-        f.write('user_pref("browser.homescreenURL", "data:text/html,mochitest-plain should start soon");\n')
+        f.write("""
+user_pref("browser.homescreenURL","app://system.gaiamobile.org");\n
+user_pref("dom.mozBrowserFramesEnabled", true);\n
+user_pref("dom.ipc.tabs.disabled", false);\n
+user_pref("dom.ipc.browser_frames.oop_by_default", true);\n
+user_pref("browser.manifestURL","app://system.gaiamobile.org/manifest.webapp");\n
+user_pref("dom.mozBrowserFramesWhitelist","app://system.gaiamobile.org,http://mochi.test:8888");\n
+user_pref("network.dns.localDomains","app://system.gaiamobile.org");\n
+""")
         f.close()
 
+        # Copy the profile to the device.
         self._dm.removeDir(self.remoteProfile)
         if self._dm.pushDir(options.profilePath, self.remoteProfile) == None:
             raise devicemanager.FileError("Unable to copy profile to device.")
+
+        # In B2G, user.js is always read from /data/local, not the profile
+        # directory.  Backup the original user.js first so we can restore it.
+        self._dm.checkCmdAs(['shell', 'rm', '-f', '%s.orig' % self.userJS])
+        if self._dm.useDDCopy:
+            self._dm.checkCmdAs(['shell', 'dd', 'if=%s' % self.userJS, 'of=%s.orig' % self.userJS])
+        else:
+            self._dm.checkCmdAs(['shell', 'cp', self.userJS, '%s.orig' % self.userJS])
+        self._dm.pushFile(os.path.join(options.profilePath, "user.js"), self.userJS)
 
         self.updateProfilesIni(self.remoteProfile)
 
@@ -355,7 +396,12 @@ def main():
     options, args = parser.parse_args()
 
     # create our Marionette instance
-    kwargs = {'emulator': options.emulator}
+    kwargs = {}
+    if options.emulator:
+        kwargs['emulator'] = options.emulator
+        auto.setEmulator(True)
+        if options.noWindow:
+            kwargs['noWindow'] = True
     if options.b2gPath:
         kwargs['homedir'] = options.b2gPath
     if options.marionette:
@@ -367,7 +413,8 @@ def main():
     auto.marionette = marionette
 
     # create the DeviceManager
-    kwargs = {'adbPath': options.adbPath}
+    kwargs = {'adbPath': options.adbPath,
+              'deviceRoot': B2GMochitest.testDir}
     if options.deviceIP:
         kwargs.update({'host': options.deviceIP,
                        'port': options.devicePort})
@@ -397,6 +444,7 @@ def main():
         retVal = mochitest.runTests(options)
     except:
         print "TEST-UNEXPECTED-FAIL | %s | Exception caught while running tests." % sys.exc_info()[1]
+        traceback.print_exc()
         mochitest.stopWebServer(options)
         mochitest.stopWebSocketServer(options)
         try:
